OSv: pin threads to vCPUs.
Normally/on Linux a process contains a single rank and thus a process is
pinned. On OSv one process runs multiple ranks, and each has to be pinned
differently. Both main and libevent worker thread has to be pinned.

Current logic to decide to which vCPU to pin is simply
my_cpu = rank % cpu_count.
Should be good (only) for most simple cases.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>