OSv: use osv_get_all_app_threads to set/get thread CPU affinity.
On Linux, CPU affinity is set between fork and exec. So we are already in
new mpi_app process, but still executing the orted code.

On OSv, we setaffinity is called from orted, and we have to explicitly
search for all threads belonging to mpi_app, and then setaffinity only
to that threads. The first thread in mpi_app is returned by osv_execve.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>