OSv fix mpirun error report on HTTP connection failure
Scenario: multiple OSv VMs are spinned up and mpirun is executed on
first of them (VM1). The very last VMs might not be fully up (they are
running, but REST API is not up yet). mpirun starts child workers in
tree manner by default - VM1 connects to say VM8 and VM8 is supposed
to connect VM16 to start mpi_app. If VM16 REST API is not up yet, then
VM8 sees TCP connection failure, but doesn't report it back to mpirun
on VM1.

This commit fixes the missing error report.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>